requirements.txt
face_alignment==1.3.5
numpy==1.20.0
opencv_python_headless==4.5.4.60
Pillow==8.4.0
scikit_image==0.18.3
skimage==0.0
tqdm==4.36.1
